move files to individual folders
