+++ src/helpers/log.py
@@ -10,7 +10,7 @@ import logging, sys
 def setLogger():
     logFormat = "%(threadName)s - %(asctime)s - %(message)s"
     logger = logging.getLogger()
-    fileHandler = logging.FileHandler("logfile.log")
+    fileHandler = logging.FileHandler("../reports/log/logfile.log")
     streamHandler = logging.StreamHandler(sys.stdout)
     formatter = logging.Formatter(logFormat, "%H:%M:%S")
     streamHandler.setFormatter(formatter)
