+++ src/model/coin/CoinQueue.py
